+++ dart/lib/src/sentry_options.dart
@@ -398,6 +398,8 @@ class SentryOptions {
   /// Enables generation of transactions and propagation of trace data. If set
   /// to null, tracing might be enabled if [tracesSampleRate] or [tracesSampler]
   /// are set.
+  @Deprecated(
+      'Use either tracesSampleRate or tracesSampler instead. This will be removed in v9')
   bool? enableTracing;
 
   /// Enables sending developer metrics to Sentry.
@@ -537,6 +539,7 @@ class SentryOptions {
   /// Returns if tracing should be enabled. If tracing is disabled, starting transactions returns
   /// [NoOpSentrySpan].
   bool isTracingEnabled() {
+    // ignore: deprecated_member_use_from_same_package
     final enable = enableTracing;
     if (enable != null) {
       return enable;